+++ src/CartaMonstruo.java
@@ -1,13 +1,13 @@
 public class CartaMonstruo extends Carta {
  
-    private int nivelCarta;
-    private int atk;
-    private int def;
-    private int boostAtk;
-    private int boostDef;
+    private byte nivelCarta;
+    private short atk;
+    private short def;
+    private short boostAtk;
+    private short boostDef;
     private boolean puedeAtacar;
  
-    public CartaMonstruo(String nombre, int nivelCarta, int atk, int def) {
+    public CartaMonstruo(String nombre, byte nivelCarta, short atk, short def) {
         super(nombre);
         this.nivelCarta = nivelCarta;
         this.atk = atk;
@@ -37,11 +37,11 @@ public class CartaMonstruo extends Carta {
         this.puedeAtacar = puedeAtacar;
     }
  
-    public void aplicarBoostAtk(int incremento) {
+    public void aplicarBoostAtk(short incremento) {
         boostAtk = incremento;
     }
  
-    public void aplicarBoostDef(int incremento) {
+    public void aplicarBoostDef(short incremento) {
         boostDef = incremento;
     }
