+++ src/EspadaDeZeus.java
@@ -1,5 +1,5 @@
 public class EspadaDeZeus extends CartaMagica {
-    private int boostAtk = 500;
+    private short boostAtk = 500;
 
     public EspadaDeZeus() {
         super("Espada de Zeus", "Aumenta +500 ATK a un monstruo en tu campo.");
